+++ src/Types.ts
@@ -1,35 +1,14 @@
+import { AcceptedPlugin, Root } from "postcss";
+
 import { PostCSSPrefixWrapOptions } from "./plugin/PostCSSPrefixWrap";
-
-export interface PostCSSInput {
-    file?: string;
-}
-
-export interface PostCSSSource {
-    input: PostCSSInput;
-}
-
-export interface PostCSSContainer {
-    source?: PostCSSSource;
-    parent?: PostCSSContainer;
-    type?: string;
-    walkRules: (callback: (atRule: PostCSSRule) => void) => void;
-}
-
-export interface PostCSSRule extends PostCSSContainer {
-    selector: string;
-}
-
-export interface PostCSSAtRule extends PostCSSContainer {
-    name: string;
-}
 
 export interface PostCSSAcceptedPlugin {
     postcssPlugin: string;
 
-    Once(root: PostCSSContainer): void;
+    Once(root: Root): void;
 }
 
-export interface PostCSS7PostCSS {
+export interface PostcssV7 {
     plugin: (
         name: string,
         thing: (
@@ -39,15 +18,8 @@ export interface PostCSS7PostCSS {
     ) => PostCSS8Plugin;
 }
 
-export interface PostCSS8PostCSS {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment,@typescript-eslint/no-explicit-any
-    // @ts-ignore
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    Root: any;
-}
-
-export type PostCSS7Plugin = () => PostCSSAcceptedPlugin;
+export type PostCSS7Plugin = () => AcceptedPlugin;
 export type PostCSS8Plugin = (
     prefixSelector: string,
     options?: PostCSSPrefixWrapOptions,
-) => PostCSSAcceptedPlugin;
+) => AcceptedPlugin;