+++ src/plugin/PostCSS7Plugin.ts
@@ -1,11 +1,11 @@
-import { PostCSS7PostCSS } from "../Types";
+import { PostcssV7 } from "../Types";
 
 import PostCSSPrefixWrap, {
     PLUGIN_NAME,
     PostCSSPrefixWrapOptions,
 } from "./PostCSSPrefixWrap";
 
-export const asPostCSSv7PluginGenerator = (postcss: PostCSS7PostCSS) => {
+export const asPostCSSv7PluginGenerator = (postcss: PostcssV7) => {
     return postcss.plugin(
         PLUGIN_NAME,
         (prefixSelector: string, options?: PostCSSPrefixWrapOptions) => {